fix: moved resource section
