+++ src/routes/LandingPage/ResourceSection/index.scss
